requirements.txt
google-generativeai
python-dotenv
youtube_transcript_api
streamlit
streamlit-modal
pypdf2
langchain
faiss-cpu
openai
huggingface_hub
InstructorEmbedding
sentence_transformers
langchain-community

langchain-openai
langchain-google-genai